Removed all static forward prototype declarations in AVSequencer with order rearrange if was necessary and more constification of the player, also changed some static function to take int parameter.

+++ libavsequencer/player.h
@@ -1978,7 +1978,7 @@ typedef struct AVSequencerPlayerEffects {
        pointed as an array with size of number of total effects.  */
     void (*effect_func)(AVSequencerContext *const avctx,
         AVSequencerPlayerHostChannel *const player_host_channel, AVSequencerPlayerChannel *const player_channel,
-        const int channel, const unsigned fx_byte, uint16_t data_word);
+        const uint16_t channel, const unsigned fx_byte, uint16_t data_word);
 
     /** Function pointer for pre-pattern evaluation. Some effects
        require a pre-initialization stage. Can be NULL if the effect
@@ -1986,14 +1986,14 @@ typedef struct AVSequencerPlayerEffects {
        pre-initialization stage.  */
     void (*pre_pattern_func)(const AVSequencerContext *const avctx,
         AVSequencerPlayerHostChannel *const player_host_channel, AVSequencerPlayerChannel *const player_channel,
-        const int channel, uint16_t data_word);
+        const uint16_t channel, uint16_t data_word);
 
     /** Function pointer for parameter checking for an effect. Can
        be NULL if the effect number either is not used or the effect
        does not require pre-checking.  */
     void (*check_fx_func)(const AVSequencerContext *const avctx,
         AVSequencerPlayerHostChannel *const player_host_channel, AVSequencerPlayerChannel *const player_channel,
-        const int channel, uint16_t *const fx_byte, uint16_t *const data_word, uint16_t *const flags);
+        const uint16_t channel, uint16_t *const fx_byte, uint16_t *const data_word, uint16_t *const flags);
 
     /** Special flags for this effect, this currently defines if the
        effect is executed during the whole row each tick or just only
